Added search

+++ src/main/java/com/example/demo/repository/DoctorRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface DoctorRepository extends CrudRepository<Doctor,Long> {
+public interface DoctorRepository extends CrudRepository<Doctor,Integer> {
 }
